fix(trusty-mpm): bring lifecycle.rs under its SLOC ratchet + fix dangling doc pointer
CI's 500-line file-size cap flagged lifecycle.rs growing 4 SLOC past its
frozen 1231-line grandfathered budget after the #3312 enforcement wiring
(the two `.await` sites the new async resolve_for_config_enforced requires).

Splits by cohesion — production route/spawn logic vs. its own test suite —
rather than reformatting around the ratchet: extracts the file's large
inline `#[cfg(test)] mod tests { ... }` (612 lines) into a companion
`lifecycle_tests.rs`, included via `#[path = "lifecycle_tests.rs"] mod
tests;`, mirroring the exact `_tests.rs` convention this codebase already
established for `gh_account_spawn_env_tests.rs`. Pure code motion: every
test is verbatim, same module path (`daemon::managed_routes::lifecycle::
tests::*`), same 15/15 passing. lifecycle.rs itself drops from 1235 to 836
SLOC; re-froze the ratchet budget down (836) via `check_line_cap.sh
--update` so it can't silently regrow. Also collapsed the two call sites'
`crate::core::git_identity::resolve_for_config_enforced` to a single
top-level `use`, avoiding a wasteful signature-wrapping detour.

Also fixes a dangling `Test:` doc-comment pointer the test-pointers.yml CI
lint caught: `resolve_for_config_enforced`'s doc cited a
`resolve_for_config_enforced_matches_plain_resolver_pre_fix_behaviour` test
that was never actually split out as its own function — the causality proof
it described lives inside `resolve_for_config_enforced_fails_closed_on_switch_failure`
already; corrected the citation to match.

Verified: cargo fmt/clippy/test all clean; check_line_cap.sh,
check_test_pointers.sh, and check_sld.sh all pass locally.

Refs #3312
